jit: silence x86-64 pointer cast warning in lookupswitch_pair_comp()
This silences a warning that surfaces on x86-64 due to casting a pointer
to an integer of a different size.

+++ jit/statement.c
@@ -152,5 +152,5 @@ int lookupswitch_pair_comp(const void *key, const void *elem)
 {
 	const struct lookupswitch_pair *pair = elem;
 
-	return (int32_t)key - pair->match;
+	return (int32_t) ((intptr_t) key) - pair->match;
 }
